[MEXEC-150] Plugin warns that "killAfter is now deprecated" for no obvious reason
This problem occures because of an oversight in the fix for MEXEC-123 in r19360, where the defalt value for killAfter was changed.
This commit now sets it back to its original value, making the warning go away. The parameter is not used so this fix will have no negative impact on users.

git-svn-id: file:///home/tiste/MOJOHAUS-TO-GIT/SVN-MOJO-WIP/trunk/mojo/exec-maven-plugin@20476 52ab4f32-60fc-0310-b215-8acea882cd1b