Fix running executableDependency at Unix-like systems
At Unix-like systems Maven installs artifacts to its local repository
without "executable" bit set. So even if the plugin dependency is an
executable binary or script, it fails to run with "permission denied"
error.

The change makes the executableDependency file executable at Unix-like
OS. At Windows it is a no-op.